bt16

+++ FuramaResort/src/model/House.java
@@ -15,7 +15,7 @@ public class House extends Services {
         this.numberOfFloors = numberOfFloors;
     }
 
-    public House(String nameServices, float useArea, int rentalCosts, int maxQuantity, String rentalType, String id, String roomStandard, String convenient, int
+    public House(String nameServices, float useArea, float rentalCosts, int maxQuantity, String rentalType, String id, String roomStandard, String convenient, int
             numberOfFloors) {
         super(nameServices, useArea, rentalCosts, maxQuantity, rentalType, id);
         this.roomStandard = roomStandard;
